Don't load undefined libeay32 foreign-library.
In commit 5860d85a8d3224b555e092b010a3bd6d7e818f31, a new foreign library
named LIBEAY32 was defined and added to the reload function.
1f2c1ed66e9ac273d5f71f40142249e989f70054 removed the LIBEAY32
foreign-library in favor of using an alternate dll name for the LIBSSL
library, but LIBEAY32 is still loaded in the RELOAD function, which causes
an "undefined foreign library" error when RELOAD is called.