Un-break LibreSSL
Commit d3e4424 addressing issue #64 caused code paths necessary for the
correct functioning of LibreSSL to be skipped. In particular:
 - Locking callbacks are no longer set, causing crashes when CL+SSL
   is used in an application with multiple threads.
 - SSL_library_init and SSL_load_error_strings are no longer called.
 - SSL_CTX_set_tmp_rsa_callback is no longer called (probably harmless,
   as this is only relevant if one is using export grade ciphersuites).

This was an easy to make oversight. The code paths above are
predicated on COMPAT-OPENSSL-VERSION being below version 1.1, since
they are no longer necessary for versions 1.1 or later. However,
LibreSSL APIs are compatible with version OpenSSL 1.0 APIs, yet the
version number is encoded as 2.0 (technically, #x20000000).

This change updates the conditional using LIBRESSLP, so that the
necessary code paths are restored. Additionally, it adds
OPENSSL-INIT-SSL and conditionally calls it instead of
SSL-INIT-LIBRARY for OpenSSL 1.1, where OPENSSL_init_ssl supersedes
SSL_library_init and SSL_load_error_strings.