Un-break LibreSSL
Commit d3e4424 addressing issue #64 caused code paths that can be
necessary for the correct functioning of OpenSSL/LibreSSL to be skipped.
In particular:
 - Locking callbacks are no longer set, causing crashes when CL+SSL
   is used in an application with multiple threads.
 - SSL_CTX_set_tmp_rsa_callback is no longer called (probably harmless,
   as this is only relevant if one is using export grade ciphersuites).
 - SSL_library_init and SSL_load_error_strings are no longer called.
   This is mostly harmless, with at least one important exception. The
   default version of LibreSSL on macOS (/usr/lib/libressl.dylib) is
   version 2.2.7. Versions prior to 2.5.2 dynamically look up some
   digests, which will fail if SSL_library_init (or other equivalents
   of OpenSSL_add_all_digests) have not previously been called. As a
   result, CL+SSL crashes on macOS, at least.

This was an easy to make oversight. The code paths above are
predicated on COMPAT-OPENSSL-VERSION being below version 1.1, since
they are no longer necessary for versions 1.1 or later. However,
LibreSSL APIs are compatible with version OpenSSL 1.0 APIs, yet the
version number is encoded as 2.0 (technically, #x20000000).

This change updates the conditional using LIBRESSLP, so that the
necessary code paths are restored.