Implement stream-fd for ABCL
This fixes for broken BIO on OpenSSL 1.1.1 ff. which causes ABCL to
segfault.

This hack allows ABCL to get at the underlying file descriptor using
undocumented portions of the JVM that are not guaranteed to be stable.

Tested under openjdk8 and openjdk11.

c.f. <https://github.com/cl-plus-ssl/cl-plus-ssl/issues/72>,
<https://abcl.org/trac/ticket/464>.