Add :canary to cffi:define-foreign-library
See
https://common-lisp.net/project/cffi/manual/cffi-manual.html#define_002dforeign_002dlibrary

This is CFFI's more general solution to pushing
`:cl+ssl-foreign-libs-already-loaded` to `*features*`. Supporting this is nice
as it requires zero work from the user.

I tried to choose symbols that appear to be present in all versions of
OpenSSL. If one of these symbols ever disappears, the canary should be updated.