SNI server-side support
Add support for handling SNI on the server. The ffi functions for working with the
SNI callback have been added, and a default callback has been defined. If `:sni t` is
passed to `make-context`, the default callback is registered. The callback can be
manually specified when calling `make-context` by specifying the `:sni-callback`
key.

When calling `make-ssl-server-stream`, a alist for SNI configuration can be passed
using the :sni-config key. The alist has the following structure:

  ((server-name-str) . ((:certificate certificate-path
                         :key key-path
                         :passphrase passphrase-str))))

If a servername is passed in the TLS handshake and the default callback has been
registered, the servername will be looked up in the sni-config. If a value is
present, the certificate and key for the SNI entry are set on the SSL object. If
no value is present, SSL_TLSEXT_ERR_NOACK is returned and the server will continue
with the :certificate and :key passed into `make-ssl-server-stream`.