Fix memory corruption and leaks on clisp
----
Files
  ffi-buffer.lisp
  ffi-buffer-clisp.lisp
  streams.lisp
  random.lisp
  x509.lisp

----
Memory corruption bug on clisp

s/b-replace
b/s-replace
had bugs that could cause them to miscalculate the buffer's end as being
beyond its boundary, or miscalculate the number of bytes to copy if the
buffer's end was specified but the sequence was smaller.
All callers of s/b-replace happened to pass arguments that didn't trigger
its bugs.
But one caller of b/s-replace (namely stream-write-sequence) could
legitimately call it in a way that did trigger one of its bugs.
E.g. if the buffer was smaller than the sequence, it would corrupt memory
by writing beyond the buffer's bounds.

I have fixed all the bugs, which were in s/b-replace and b/s-replace.

----
Performance

For clisp:

b/s-replace also copies less.
The old version always called subseq, which copies.
The new version copies only if the source seq is not a vector.

s/b-replace is not expected to allocate memory proportional to its
input arrays.
But due to its call to memory-as, it did.
Now it doesn't. It allocates O(1) memory, regardless of input array sizes.

b/s-replace also allocates O(1) memory now.

For all lisps:

stream-read-sequence
I have made this clearer.

stream-write-sequence
I have rewritten this to be clearer.
And faster. The old version could flush a non-full stream.
As a pathological case,
  writing 1 byte, then 2048 times writing 2049 bytes, would cause 4096
  flushes.
Now that will cause only 2049 flushes.

----
Memory leak on clisp

There was also a memory leak because foreign buffers were allocated but
never freed. I have fixed this by extending the buffer API and having
all callers of make-buffer also release the buffer when finished with it.