Stabilise performance for byte-sized operations
stream-listen
stream-read-byte
are written to potentially have horrible performance.

The reason is that cffi's experimental Shareable Byte Vector interface
is bad. It indicates that although with-pointer-to-vector-data (wptvd)
with an empty body may be constant time (due to the vector's being shared
between Lisp and C), it may also be O(len(vector)) on implementations
that require copying in/out. Given that cffi is a portability layer,
the name make-shareable-byte-vector is misleading if the result isn't
necessarily shared between C and Lisp but may require copying. Anyway,
bad cffi names aside, callers must therefore treat wptvd as having
O(len(vector)) overhead since that is what its documentation indicates.

In that light, stream-listen and stream-read-byte, which are written to
potentially cause the copying of the entire underlying buffer when they
only want to operate on one byte, are badly written. They are written
with a mistaken assumption, that wptvd has O(1) overhead, but it is
really specified by cffi to have O(len(vector)) overhead.

One solution is for the stream to own another one byte input buffer for
these two functions to use. The input buffer is only used as a temporary
variable within a function. It doesn't retain meaningful state between
function calls. So this solution is easy to implement.

As it happens, with the clisp-specific ffi-buffer-clisp.lisp,
the clisp wptvd in cl+ssl is O(1) so there happens not to be a
problem presently. All current wptvd's in use by cl+ssl are O(1). But
stream-listen and stream-read-byte should still not be written to assume
that the general wptvd has O(1) overhead. A new wptvd may be added
to cffi or an old one rewritten to have the allowable O(len(vector))
overhead. Thus stream-listen and stream-read-byte should really be
corrected to reflect that wptvd does not guarantee that implementations
will have less than O(len(vector)) overhead. As written, they are just
lucky they happen to perform well because of the implementation details
of the current wptvd's in existence.

This commit implements the solution of having a separate small buffer
for stream-listen and stream-read-byte to use.

Performance will not improve today. But this design will ensure that
cl+ssl doesn't in the future suddenly mysteriously develop horrible
performance due to relying on implementation details of wptvd adding O(1)
overhead when it is really specified to add O(len(vector)) overhead.