Remove broken defcfun-late-bound macro
The macro is only called in cmucl. It is written to raise an error if
called for vararg functions. But ERR_set_error in the same file is a
vararg function that results in the macro's being called. Thus the file
cannot be compiled/loaded in cmucl.

In the future if other Lisps also require the macro to be called, the file
will not be loadable on those Lisps either, for the same reason as above.

Further, this macro was created to solve a problem with cmucl's compiler
being too optimising and complaining if a foreign function didn't exist
in the foreign library.
https://gitlab.common-lisp.net/cmucl/cmucl/-/issues/74
The fix described there was only partly implemented by cffi at the time,
so this macro was necessary. But now cffi has implemented it completely
(in revision a068a492892b12d54b430d93d707d663da850210) so this macro
isn't necessary.